mybackend/todos/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.7
aspy.refactor-imports==2.1.1
attrs==19.3.0
black==19.10b0
cached-property==1.5.1
certifi==2020.4.5.1
cfgv==3.1.0
click==7.1.2
distlib==0.3.0
Django==3.0.6
django-model-utils==4.0.0
filelock==3.0.12
flake8==3.8.2
flake8-bugbear==20.1.4
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
identify==1.4.18
isort==4.3.21
mccabe==0.6.1
nodeenv==1.3.5
pathspec==0.8.0
pre-commit==2.4.0
promise==2.3
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
Rx==1.6.1
seed-isort-config==2.1.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.21
